remove flake8 errors
